requirements.txt
flask
python-dotenv
flask-talisman
flask-limiter
flask-wtf
authlib
pyyaml
markdown
gunicorn
requests
pymongo
dnspython
